Add placeholder Home page.
