+++ lib/pages/home.dart
@@ -1,8 +1,5 @@
 import 'package:flutter/material.dart';
 
-import 'package:tasty_tracks/pages/search/search.dart';
-import 'package:tasty_tracks/pages/settings.dart';
-
 class HomePage extends StatefulWidget {
   static final String routeName = '/home';
   final String pageTitle = 'Home';
@@ -12,44 +9,22 @@ class HomePage extends StatefulWidget {
 }
 
 class _HomePageState extends State<HomePage> {
-  int _selectedIndex = 0;
-  final _pages = <Widget>[
-    Text('TODO Home page'),
-    SearchPage(),
-    SettingsPage(),
-  ];
-
-  final _items = <BottomNavigationBarItem>[
-    BottomNavigationBarItem(
-        icon: Icon(Icons.home), title: Text('Home')
-    ),
-    BottomNavigationBarItem(
-        icon: Icon(Icons.search), title: Text('Search')
-    ),
-    BottomNavigationBarItem(
-        icon: Icon(Icons.settings), title: Text('Settings')
-    ),
-  ];
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: _currentPage(),
-      bottomNavigationBar: BottomNavigationBar(
-        items: _items,
-        currentIndex: _selectedIndex,
-        onTap: _onItemTapped,
+      appBar: AppBar(
+        title: Text(widget.pageTitle),
+      ),
+      body: SafeArea(
+        child: Center(
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              Text('Home'),
+            ],
+          ),
+        ),
       ),
     );
-  }
-
-  Widget _currentPage() {
-    return _pages[_selectedIndex];
-  }
-
-  void _onItemTapped(int index) {
-    setState(() {
-      _selectedIndex = index;
-    });
   }
 }
